agents/requirements.txt
google-adk[extensions]>=0.1.0
fastapi>=0.111.0
uvicorn>=0.30.1
python-dotenv>=1.0.1
httpx>=0.27.0
pydantic>=2.7.0
